maa is closed

+++ libft/ft_strlen.c
@@ -6,15 +6,15 @@
 /*   By: seunlee2 <seunlee2@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/03/17 17:38:15 by seunlee2          #+#    #+#             */
-/*   Updated: 2023/03/24 16:01:14 by seunlee2         ###   ########.fr       */
+/*   Updated: 2023/08/25 17:08:20 by seunlee2         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-size_t	ft_strlen(const char *s)
+int	ft_strlen(const char *s)
 {
-	size_t	len;
+	int	len;
 
 	len = 0;
 	while (s[len])
